Remove OwnProps for Assessment and Workspace
